requirements.txt
streamlit==1.32.2
tensorflow==2.15.0
numpy==1.26.4
pillow==10.2.0
opencv-python-headless==4.9.0.80